// frontend/src/api/generated/model/otaCreateOTAPackageRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */

export interface OtaCreateOTAPackageRequest {
  checksum?: string;
  description?: string;
  fileSize?: number;
  fileUrl?: string;
  packageName: string;
  packageType?: string;
  product_key: string;
  releaseNotes?: string;
  status?: string;
  uploadType?: string;
  version?: string;
}